+++ components/organisms/notPwaHeader.tsx
@@ -6,7 +6,7 @@ declare global {
   }
 }
 
-import React,{useState,useEffect} from 'react';
+import React, {useState,useEffect } from 'react';
 import AppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
