+++ features/api/petStatDecrease.ts
